backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
supabase>=2.0.0
openai>=1.50.0
httpx>=0.27.0
tenacity>=8.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
